refactor: move log_utils to utils module
Move log_utils.py to utils/log_utils.py and update all imports to reflect new location. Update __init__.py files to maintain proper module exports.
