+++ src/mcp_coder/__init__.py
@@ -28,7 +28,6 @@ from .llm_providers.claude.claude_executable_finder import (
     find_claude_executable,
     verify_claude_installation,
 )
-from .log_utils import setup_logging
 from .prompt_manager import (
     get_prompt,
     validate_prompt_directory,
@@ -78,6 +77,4 @@ __all__ = [
     "get_prompt",
     "validate_prompt_markdown",
     "validate_prompt_directory",
-    # Logging utilities
-    "setup_logging",
 ]
