+++ src/mcp_coder/utils/__init__.py
@@ -24,6 +24,7 @@ from .git_operations import (
     stage_all_changes,
     stage_specific_files,
 )
+from .log_utils import log_function_call, setup_logging
 from .subprocess_runner import (
     CommandOptions,
     CommandResult,
@@ -44,6 +45,9 @@ __all__ = [
     "is_git_repository",
     "stage_all_changes",
     "stage_specific_files",
+    # Logging utilities
+    "log_function_call",
+    "setup_logging",
     # Subprocess operations
     "CommandOptions",
     "CommandResult",
