+++ tests/utils/test_log_utils.py
@@ -5,12 +5,11 @@ import logging
 import os
 import tempfile
 from pathlib import Path
-from unittest import mock
 from unittest.mock import MagicMock, patch
 
 import pytest
 
-from mcp_coder.log_utils import log_function_call, setup_logging
+from mcp_coder.utils.log_utils import log_function_call, setup_logging
 
 
 class TestSetupLogging:
@@ -81,7 +80,7 @@ class TestSetupLogging:
 class TestLogFunctionCall:
     """Tests for the log_function_call decorator."""
 
-    @patch("mcp_coder.log_utils.stdlogger")
+    @patch("mcp_coder.utils.log_utils.stdlogger")
     def test_log_function_call_basic(self, mock_stdlogger: MagicMock) -> None:
         """Test the basic functionality of the decorator."""
 
@@ -97,7 +96,7 @@ class TestLogFunctionCall:
         assert result == 3
         assert mock_stdlogger.debug.call_count == 2  # Called for start and end logging
 
-    @patch("mcp_coder.log_utils.stdlogger")
+    @patch("mcp_coder.utils.log_utils.stdlogger")
     def test_log_function_call_with_exception(self, mock_stdlogger: MagicMock) -> None:
         """Test that exceptions are properly logged."""
 
@@ -114,7 +113,7 @@ class TestLogFunctionCall:
         assert mock_stdlogger.debug.call_count == 1
         assert mock_stdlogger.error.call_count == 1
 
-    @patch("mcp_coder.log_utils.stdlogger")
+    @patch("mcp_coder.utils.log_utils.stdlogger")
     def test_log_function_call_with_path_param(self, mock_stdlogger: MagicMock) -> None:
         """Test that Path objects are properly serialized."""
 
@@ -159,7 +158,7 @@ class TestLogFunctionCall:
             test_path
         ) in str(result_arg)
 
-    @patch("mcp_coder.log_utils.stdlogger")
+    @patch("mcp_coder.utils.log_utils.stdlogger")
     def test_log_function_call_with_large_result(
         self, mock_stdlogger: MagicMock
     ) -> None:
@@ -186,8 +185,8 @@ class TestLogFunctionCall:
         result_arg = second_call[0][3]
         assert "<Large result of type list" in result_arg
 
-    @patch("mcp_coder.log_utils.structlog")
-    @patch("mcp_coder.log_utils.stdlogger")
+    @patch("mcp_coder.utils.log_utils.structlog")
+    @patch("mcp_coder.utils.log_utils.stdlogger")
     def test_log_function_call_with_structured_logging(
         self, mock_stdlogger: MagicMock, mock_structlog: MagicMock
     ) -> None:
@@ -196,7 +195,7 @@ class TestLogFunctionCall:
         mock_structlogger = mock_structlog.get_logger.return_value
 
         # Mock to simulate FileHandler being present
-        with patch("mcp_coder.log_utils.any", return_value=True):
+        with patch("mcp_coder.utils.log_utils.any", return_value=True):
             # Define a test function
             @log_function_call
             def test_func(a: int, b: int) -> int:
